+++ apps/example-app/src/main.ts
@@ -1,7 +1,7 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import { ExampleAppModule } from './app/example-app.module';
 import { environment } from './environments/environment';
 import {registerWindowErrorHandler} from "@senstate/client";
 import {setSenstateConnection} from "@senstate/client-connection";
@@ -17,5 +17,5 @@ if (environment.production) {
 }
 
 platformBrowserDynamic()
-  .bootstrapModule(AppModule)
+  .bootstrapModule(ExampleAppModule)
   .catch(err => console.error(err));