+++ apps/senstate-dashboard/src/app/app.component.ts
@@ -1,4 +1,4 @@
-import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {ChangeDetectionStrategy, Component, TrackByFunction} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {map} from 'rxjs/operators';
 import {App, DASHBOARD_EVENT_NAMES, NetworkInfo} from '@senstate/dashboard-connection';
@@ -8,6 +8,8 @@ import {HubService} from "./state/hub.service";
 import {MatSliderChange} from "@angular/material/slider";
 import {MaterialCssVarsService} from "angular-material-css-vars";
 import {SocketService} from "./services/socket.service";
+import { version } from '../../../senstate/src/root-assets/package.json';
+import {DebugToggleService} from "./services/debug-toggle.service";
 
 @Component({
   selector: 'senstate-root',
@@ -30,11 +32,18 @@ export class AppComponent {
 
   darkTheme = true;
 
+  version = version;
+
+  public trackByAppFunc: TrackByFunction<App> = (index, item) => {
+    return item.appId;
+  };
+
   constructor (private http: HttpClient,
                private socketService: SocketService,
                private hubService: HubService,
                private route: ActivatedRoute,
-               public materialCssVarsService: MaterialCssVarsService) {
+               public materialCssVarsService: MaterialCssVarsService,
+               public debugToggle: DebugToggleService) {
     this.route.queryParamMap.subscribe(q => {
       this.isMobile = q.has('mobile');
       console.info('params', q)
@@ -56,10 +65,6 @@ export class AppComponent {
 
       this.hubService.statusChanged(value);
     });
-  }
-
-  public trackByAppFunc (appObj: App) {
-    return appObj.appId;
   }
 
   changeDebounce ($event: MatSliderChange) {