+++ apps/senstate-dashboard/src/app/components/shared/group-holder/group-holder.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+  selector: 'senstate-group-holder',
+  templateUrl: './group-holder.component.html',
+  styleUrls: ['./group-holder.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class GroupHolderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}