+++ apps/senstate-dashboard/src/app/components/shared/shared.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { GroupHolderComponent } from './group-holder/group-holder.component';
+
+
+
+@NgModule({
+    declarations: [GroupHolderComponent],
+    exports: [
+        GroupHolderComponent
+    ],
+    imports: [
+        CommonModule
+    ]
+})
+export class SharedModule { }