+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-masonry/watchers-masonry.component.spec.ts
@@ -22,6 +22,7 @@ describe('WatchersMasonryComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(WatchersMasonryComponent);
     HubServiceMock.getLatest().getWatchersByApp$ = appId => NEVER;
+    HubServiceMock.getLatest().getGroupedWatchersByApp$ = appId => NEVER;
     component = fixture.componentInstance;
     fixture.detectChanges();
   });