+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-masonry/watchers-masonry.module.ts
@@ -13,6 +13,8 @@ import {MatIconModule} from "@angular/material/icon";
 import {MatButtonModule} from "@angular/material/button";
 import {MatCheckboxModule} from "@angular/material/checkbox";
 import {WatchHistoryComponent} from './watchers/watch-history/watch-history.component';
+import {WatchDiffComponent} from "./watchers/watch-diff/watch-diff.component";
+import {SharedModule} from "../../shared/shared.module";
 
 
 @NgModule({
@@ -23,22 +25,24 @@ import {WatchHistoryComponent} from './watchers/watch-history/watch-history.comp
     JsonWatcherComponent,
     WatcherCardComponent,
     WatchHistoryComponent,
-    HistoryTemplateDirective
+    HistoryTemplateDirective,
+    WatchDiffComponent
   ],
   exports: [
     WatchersMasonryComponent,
   ],
-  imports: [
-    CommonModule,
+    imports: [
+        CommonModule,
 
-    NgxMasonryModule,
-    MatCardModule,
-    JsonViewerModule,
-    MatMenuModule,
-    MatIconModule,
-    MatButtonModule,
-    MatCheckboxModule
-  ]
+        NgxMasonryModule,
+        MatCardModule,
+        JsonViewerModule,
+        MatMenuModule,
+        MatIconModule,
+        MatButtonModule,
+        MatCheckboxModule,
+        SharedModule
+    ]
 })
 export class WatchersMasonryModule {
 }